App on Undisposed Garbage Notification for High School Hackathon
This application was made in my 8th grade for a high school hackathon using the MIT App Inventor tool. With garbage thrown all across roads and public areas in my city, Bengaluru, and with no action being taken, I wanted to put my efforts towards the app. The App has a User Registration and Login system with OTP connected to a database. The user can notify any garbage dump in the locality by just taking a photo of it through the application. The application scans the location details and automatically notifies the local administrative body of the details to take adequate action. The details can also be notified to any other individual to take action. The users can also view the areas with undisposed garbage. The application was well received by my teachers and we were one of the winners of the event.