Checking for Vampire number
Derived my own mathematical logic to check if a number is a vampire number or not. (In number theory, a vampire number is a composite natural number with an even number of digits, that can be factored into two natural numbers each with half as many digits as the original number, where the two factors contain precisely all the digits of the original number, in any order, counting multiplicity). Done during my free time in Class 11.