Computer playing the Othello game against user
Derived a mathematical logic to make the computer play the Othello Game. The algorithm checks for the largest number of flips it can make in one move along all directions. It also swaps chances if there is no move for any side to play. A counter shows the number of coins of each color present and at the end determines the winner. 
Done in Grade 10 during free time.